Refactor `IOAppSpec`
