+++ tests/js/src/main/scala/catseffect/examplesplatform.scala
@@ -106,19 +106,4 @@ package examples {
     def run(args: List[String]): IO[ExitCode] = IO.pure(ExitCode.Success)
   }
 
-  // stub
-  object EvalOnMainThread extends IOApp {
-    def run(args: List[String]): IO[ExitCode] = IO.never
-  }
-
-  // stub
-  object MainThreadReportFailure extends IOApp {
-    def run(args: List[String]): IO[ExitCode] = IO.never
-  }
-
-  // stub
-  object BlockedThreads extends IOApp.Simple {
-    val run = IO.never
-  }
-
 }
